🚑 rename: 마이페이지 파일 구조 및 이름 변경
